move scneario implementation in related context
